% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryFxn.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Extracting a certain country population stats and producing a graph.}
\usage{
CountryPopulation(country)
}
\arguments{
\item{country}{A character string}
}
\value{
A graph produced from the \code{country} statistics
}
\description{
This function will take in a string that will be of a country within the data
set and extract that row from the data and produce a graph that displays how
the population of that area has increased/decreased or stayed stagnant
through out the time lapse of 1950-2020.
}
\examples{
CountryPopulation('Italy')
}
