% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldCup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World Cup Data}
\format{
A data frame with 22 obs. with 4 columns.
\describe{
   \item{Attendance}{Provides the number of attendees}
   \item{Matches}{The number of matches during that World Cup}
   \item{Average}{ Provides the average number of attendees }
   \items{WorldCup}{ The location and the year the games were hosted}
}
}
\usage{
World_Cup
}
\description{
In this data set we can see the statistics of the world cup years and
 locations they were hosted. As well as the amount of matches and attendees.
}
\keyword{datasets}
