% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulationCleanData.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population Data}
\format{
A data frame with 16685 with 3 columns.
\describe{
   \item{Country}{Provides the country name}
   \item{Year}{Provides the year that the statistic is from}
   \item{Population}{ Provides the number of humans within that area }
}
}
\usage{
WorldPopulation
}
\description{
In this data set we can see the statistics of the world populations
 per country and year.
}
\keyword{datasets}
